+++ src/linear_model/sgdRegressor.ts
@@ -32,8 +32,9 @@ import {
   convertToNumericTensor1D_2D,
   convertToNumericTensor2D
 } from '../utils'
-import { Scikit2D, Scikit1D } from '../types'
+import { Scikit2D, Scikit1D, OptimizerTypes, LossTypes } from '../types'
 import { RegressorMixin } from '../mixins'
+import { fromJson, toJSON } from './modelSerializer'
 /**
  * SGD is a thin Wrapper around Tensorflow's model api with a single dense layer.
  * With this base class and different error functions / regularizers we can
@@ -91,6 +92,10 @@ export interface SGDRegressorParams {
    */
 
   isClassification?: boolean
+
+  optimizerType: OptimizerTypes
+
+  lossType: LossTypes
 }
 
 export class SGDRegressor extends RegressorMixin {
@@ -99,11 +104,15 @@ export class SGDRegressor extends RegressorMixin {
   modelCompileArgs: ModelCompileArgs
   denseLayerArgs: DenseLayerArgs
   isMultiOutput: boolean
+  optimizerType: OptimizerTypes
+  lossType: LossTypes
 
   constructor({
     modelFitArgs,
     modelCompileArgs,
-    denseLayerArgs
+    denseLayerArgs,
+    optimizerType,
+    lossType
   }: SGDRegressorParams) {
     super()
     this.model = sequential()
@@ -111,6 +120,8 @@ export class SGDRegressor extends RegressorMixin {
     this.modelCompileArgs = modelCompileArgs
     this.denseLayerArgs = denseLayerArgs
     this.isMultiOutput = false
+    this.optimizerType = optimizerType
+    this.lossType = lossType
   }
 
   /**
@@ -246,7 +257,9 @@ export class SGDRegressor extends RegressorMixin {
     return {
       modelFitArgs: this.modelFitArgs,
       modelCompileArgs: this.modelCompileArgs,
-      denseLayerArgs: this.denseLayerArgs
+      denseLayerArgs: this.denseLayerArgs,
+      optimizerType: this.optimizerType,
+      lossType: this.lossType
     }
   }
 
@@ -380,5 +393,14 @@ export class SGDRegressor extends RegressorMixin {
     }
 
     return intercept
+  }
+
+  public async toJson(): Promise<string> {
+    const classifierJson = JSON.parse(super.toJson() as string)
+    return toJSON(this, classifierJson)
+  }
+
+  public fromJson(model: string) {
+    return fromJson(this, model) as this
   }
 }