+++ src/mixins.ts
@@ -1,8 +1,8 @@
 import { Tensor1D, Tensor2D } from '@tensorflow/tfjs-core'
 import { Scikit2D, Scikit1D } from './types'
 import { r2Score, accuracyScore } from './metrics/metrics'
-
-export class TransformerMixin {
+import Serialize from './serialize'
+export class TransformerMixin extends Serialize {
   // We assume that fit and transform exist
   [x: string]: any
 
@@ -20,7 +20,7 @@ export class PredictorMixin {
   }
 }
 
-export class RegressorMixin {
+export class RegressorMixin extends Serialize {
   // We assume we have a predict function
   [x: string]: any
 
@@ -31,7 +31,7 @@ export class RegressorMixin {
   }
 }
 
-export class ClassifierMixin {
+export class ClassifierMixin extends Serialize {
   // We assume we have a predict function
   [x: string]: any