+++ src/naive_bayes/baseNaiveBayes.ts
@@ -17,6 +17,7 @@ import { tf } from '../shared/globals'
 import { Scikit1D, Scikit2D } from '../types'
 import { convertToNumericTensor2D, convertToTensor1D } from '../utils'
 import { Tensor1D } from '@tensorflow/tfjs-core'
+import Serialize from '../serialize'
 
 export interface NaiveBayesParams {
   /**
@@ -31,7 +32,7 @@ export interface NaiveBayesParams {
   varSmoothing?: number
 }
 
-export abstract class BaseNaiveBayes {
+export abstract class BaseNaiveBayes extends Serialize {
   priors?: tf.Tensor1D
   varSmoothing: number
 
@@ -40,6 +41,7 @@ export abstract class BaseNaiveBayes {
   public variances: tf.Tensor1D[]
 
   constructor(params: NaiveBayesParams = {}) {
+    super()
     this.classes = tf.tensor1d([])
     this.means = []
     this.variances = []
@@ -151,4 +153,37 @@ export abstract class BaseNaiveBayes {
     mean: tf.Tensor1D,
     variance: tf.Tensor1D
   ): tf.Tensor1D
+
+  public toJson(): string {
+    const jsonClass = JSON.parse(super.toJson() as string)
+
+    if (this.priors) {
+      jsonClass.priors = this.priors.arraySync()
+    }
+    jsonClass.classes = this.classes.arraySync()
+    jsonClass.means = this.means.map((t: Tensor1D) => t.arraySync())
+    jsonClass.variances = this.variances.map((v: Tensor1D) => v.arraySync())
+    return JSON.stringify(jsonClass)
+  }
+
+  public fromJson(model: string) {
+    const jsonModel = JSON.parse(model)
+
+    if (jsonModel.priors) {
+      jsonModel.priors = tf.tensor(jsonModel.priors)
+    }
+    jsonModel.classes = tf.tensor(jsonModel.classes)
+
+    const means = []
+    for (const wMeans of jsonModel.means) {
+      means.push(tf.tensor(wMeans))
+    }
+    const variances = []
+    for (const variance of jsonModel.variances) {
+      variances.push(tf.tensor(variance))
+    }
+    jsonModel.means = means
+    jsonModel.variances = variances
+    return Object.assign(this, jsonModel) as this
+  }
 }